# docs/swagger.yaml
definitions:
  reqmodel.CreateDiscount:
    properties:
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
      type:
        type: string
      value:
        type: number
    required:
    - end_date
    - name
    - start_date
    - type
    - value
    type: object
  reqmodel.CreateDiscountProductTarget:
    properties:
      discount_id:
        type: string
      is_active:
        type: boolean
      max_total_quota:
        type: integer
      price_after_discount:
        type: number
      price_before_discount:
        type: number
      product_name:
        type: string
      target_id:
        type: string
      total_discount:
        type: number
    required:
    - discount_id
    - is_active
    - target_id
    type: object
...
  reqmodel.UpdateDiscount:
    properties:
      end_date:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
  reqmodel.UpdateDiscountProductTarget:
    properties:
      discount_id:
        type: string
      is_active:
        type: boolean
      max_total_quota:
        type: integer
      price_after_discount:
        type: number
      price_before_discount:
        type: number
      product_name:
        type: string
      target_id:
        type: string
      total_discount:
        type: number
    required:
    - discount_id
    - is_active
    - target_id
    type: object
...
  resmodel.DataResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  resmodel.DatasResponse:
    properties:
      datas:
        items: {}
        type: array
      message:
        type: string
      pagination: {}
      success:
        type: boolean
    type: object
  resmodel.NoDataResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.DataValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Invalid email address
        type: string
    type: object
  utils.RequestError:
    properties:
      fields:
        items:
          $ref: '#/definitions/utils.DataValidationError'
        type: array
      message:
        example: Invalid email address
        type: string
      status_code:
        example: 422
        type: integer
    type: object
info:
  contact:
    email: lifelinejar@mail.com
    name: API Support
  description: Service Discount Rest API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Service Discount
  version: "1.0"
paths:
  /ndi/discount:
    get:
      description: Get all discounts
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page Number
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Discounts retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DatasResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Discounts not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: List Discount
      tags:
      - Discount
    post:
      consumes:
      - application/json
      description: Create a new discount
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Create discount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.CreateDiscount'
      produces:
      - application/json
      responses:
        "201":
          description: Discount created
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Create Discount
      tags:
      - Discount
  /ndi/discount-product:
    get:
      description: Get all discount product targets
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page Number
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Discount product targets retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DatasResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Discount product targets not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: List Discount Product Target
      tags:
      - Discount Product Target
    post:
      consumes:
      - application/json
      description: Create a new discount product target
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Create discount product target
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.CreateDiscountProductTarget'
      produces:
      - application/json
      responses:
        "201":
          description: Discount product target created
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Create Discount Product Target
      tags:
      - Discount Product Target
  /ndi/discount-product/{id_unik}:
    delete:
      consumes:
      - application/json
      description: Delete a discount product target
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Discount Product Target ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount product target deleted
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Delete Discount Product Target
      tags:
      - Discount Product Target
    get:
      description: Get a discount product target by its ID
      parameters:
      - description: Discount Product Target ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount product target retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Discount product target not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Get Discount Product Target By ID
      tags:
      - Discount Product Target
    put:
      consumes:
      - application/json
      description: Update an existing discount product target
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Discount Product Target ID
        in: path
        name: id_unik
        required: true
        type: string
      - description: Update discount product target
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.UpdateDiscountProductTarget'
      produces:
      - application/json
      responses:
        "200":
          description: Discount product target updated
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Update Discount Product Target
      tags:
      - Discount Product Target
...
  /ndi/discount/{id_unik}:
    delete:
      consumes:
      - application/json
      description: Delete a discount
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount deleted
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Delete Discount
      tags:
      - Discount
    get:
      description: Get a discount by its ID
      parameters:
      - description: Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Discount not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Get Discount By ID
      tags:
      - Discount
    put:
      consumes:
      - application/json
      description: Update an existing discount
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      - description: Update discount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.UpdateDiscount'
      produces:
      - application/json
      responses:
        "200":
          description: Discount updated
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Update Discount
      tags:
      - Discount
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"